Tune-up chunking (#47)
* update chunking strategy
  - smaller 3d chunks for bigstitcher
  - moves rechunking from zarr2bdv to apply_flatfield_corr
* make logging visible for bigstitcher (for coiled debugging)

+++ workflow/scripts/zarr_to_n5_bdv.py
@@ -90,6 +90,7 @@ update_xml_h5_to_n5(snakemake.params.temp_xml,snakemake.output.bdv_xml,snakemake
 print('removing empty bdv h5/xml')
 rmtree(temp_bdv_dir)
 
+
 print('writing data to n5')
 n5_store = zarr.n5.N5Store(snakemake.output.bdv_n5)
 
@@ -98,7 +99,7 @@ for setup_i,((chan_i,chan),(tile_i,tile)) in enumerate(product(enumerate(metadat
     ds_list=[] #for setup-level attrs
     for ds in range(max_downsampling_layers):
         step=2**ds  #1,2,4,8.. 
-        zstack = da.squeeze(darr[tile_i,chan_i,:,::step,::step])
+        zstack = da.squeeze(darr[tile_i,chan_i,:,::step,::step]).astype(np.int16)
         print(f'writing to setup{setup_i}/timepoint0/s{ds}')
         with ProgressBar():
             zstack.to_zarr(n5_store,component=f'setup{setup_i}/timepoint0/s{ds}',overwrite=True,compute=True)
